Completed Chapter 1.
Not sure how this repo will go. There isn't much in the way of problem
solving, just typing out the examples. Getting them to run, however, is
not trivial and will only get more complicated.

Also, some examples do NOT use the 'best' available language features so
there are opportunities to change some of the sources.